Create a Utils class to parse LangString and Vocabulary data.
A new class (Educa\DSB\Client\Utils) will handle the parsing of
LangString data. It will replace the logic used in the LomDescription
class. Furthermore, this class will also provide a method for
extracting the name of a Vocabulary entry. If the Vocabulary entry
contains Ontology data, this will be parsed and treated as a
LangString (which it is). If not, it will simply return the raw
"name" key, which it the Vocabulary entry's unique identifier.