Standardize PER term identifiers
The usage of parent-prefixes for term identifiers was not standard, and
is now deprecated. Instead, the same identifiers as provided by the
BDPER API should be used. This causes several issues, like the inability
for the PER to be used/parsed effectively as a taxonomy path, which is
why this piece of the logic has been removed (including unit tests).

This also has the impact that it's no longer up to the BaseCurriculum to
decide if new or stale data should be used when using the termFactory()
method. Instead, it simply passes the exact information recieved, and
it's up to the implementing sub-classes to make the call if the data
should be overwritten with more up-to-date data.

Normalize term types as well, basing them on the path components of the
BDPER API. For example, because data for an objective is found at
/objectifs/:id, the type for objectives is now "objectifs" (plural).