Add support for the new LOM-CH format; deprecate getLomId()
The new LOM-CHv1.2 format more strictly adheres to the LOM format,
meaning the general.description field is now a list. Add a new method to
check whether we are dealing with a legacy format, and adapt the
getDescription() method accordingly.

Deprecate the getLomId() method. In a future release of the dsb REST
API, the lomId will not be returned as part of the payload anymore, but
will be sent as HTTP headers. Put up a warning when using the method,
and mark it as deprecated in the LomDescriptionInterface.